refactor: use the correct components in tasks screens
