+++ app/src/main/java/com/moscow/tudee/presentation/component/home_components/TaskList.kt
@@ -22,7 +22,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.moscow.tudee.R
 import com.moscow.tudee.domain.entity.Task.Status
-import com.moscow.tudee.presentation.category.toCategoryUi
+import com.moscow.tudee.presentation.screen.category.toCategoryUi
 import com.moscow.tudee.presentation.component.TudeeText
 import com.moscow.tudee.presentation.designSystem.component.PriorityChip
 import com.moscow.tudee.presentation.designSystem.component.TaskCard
